+++ changelog.md
@@ -6,6 +6,8 @@
 
 - `StickyMittenAvatarController`:
   - For scenes `2a`, `2b`, and `2c`, added layouts `1` and `2`
+- `StaticObjectInfo`:
+  - Added: `container` True if this object is a container.
 
 ### Backend